# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VTEI Student API",
        "description": "REST API для мобільного додатку студентів ВТЕІ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/disciplines/electives": {
            "get": {
                "tags": [
                    "Disciplines"
                ],
                "summary": "Список вибіркових дисциплін",
                "description": "Повертає доступні вибіркові дисципліни для освітньої програми і курсу студента. Включає інформацію про поточну/наступну хвилю вибору та кількість студентів що обрали кожну дисципліну.",
                "operationId": "17ff036916753913563d91a6d5cbd68d",
                "responses": {
                    "200": {
                        "description": "Список вибіркових дисциплін з інфо про хвилю"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/disciplines/electives/selected": {
            "get": {
                "tags": [
                    "Disciplines"
                ],
                "summary": "Обрані вибіркові дисципліни",
                "description": "Повертає дисципліни, які студент обрав у поточному навчальному році.",
                "operationId": "c426d28f90e7f0ad94bd650f150c666c",
                "responses": {
                    "200": {
                        "description": "Обрані дисципліни"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/disciplines/electives/save": {
            "post": {
                "tags": [
                    "Disciplines"
                ],
                "summary": "Зберегти вибір дисциплін",
                "description": "Зберігає вибрані студентом вибіркові дисципліни. Перезаписує поточний вибір для вказаного семестру.",
                "operationId": "29f7c2d446ade77a82bb7edf1192f92a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "predmet_ids",
                                    "sem"
                                ],
                                "properties": {
                                    "predmet_ids": {
                                        "description": "ID дисциплін",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "sem": {
                                        "description": "Семестр (1 або 2)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Вибір збережено"
                    },
                    "400": {
                        "description": "Вибір зараз недоступний"
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/disciplines/electives/{id}": {
            "delete": {
                "tags": [
                    "Disciplines"
                ],
                "summary": "Видалити вибіркову дисципліну",
                "description": "Видаляє одну дисципліну з вибору студента.",
                "operationId": "d0cc1ceb5889fe8f0034b958a52af0f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID запису з nv_predmet_student_stage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Видалено"
                    },
                    "404": {
                        "description": "Не знайдено"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/disciplines/history": {
            "get": {
                "tags": [
                    "Disciplines"
                ],
                "summary": "Історія дисциплін",
                "description": "Всі дисципліни студента по навчальних роках (обов'язкові + вибіркові).",
                "operationId": "d8392ecf096564543d086bbfb59debcc",
                "responses": {
                    "200": {
                        "description": "Дисципліни по роках"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}